requirements.txt
sentence-transformers>=2.6.0
scikit-learn>=1.3
numpy
pandas
gradio>=4.0
pypdf>=4.0
python-docx
nltk
regex
tesseract 
poppler
